refactor: split vsock host process helpers
